requirements.txt
pandas
numpy
requests
pyyaml
matplotlib
tqdm
finance-datareader
python-dateutil
flask
flask-cors
websockets
